requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
seaborn
reportlab
pyyaml
lxml
openpyxl
groq
python-dotenv
duckdb
polars